Move the train-test helper into ETL
