+++ irvine/air_quality/aq_models.py
@@ -3,18 +3,16 @@
 from collections.abc import Callable
 
 import numpy as np
-import pandas as pd
 from beartype import beartype
 from numpy.typing import NDArray
 from sklearn.ensemble import HistGradientBoostingRegressor, RandomForestRegressor
 from sklearn.linear_model import ElasticNet, LinearRegression
 from sklearn.metrics import mean_squared_error
-from sklearn.model_selection import train_test_split
 from sklearn.neighbors import KNeighborsRegressor
 from sklearn.preprocessing import StandardScaler
 from sklearn.svm import SVR
 
-from irvine.air_quality.aq_etl import find_derivatives, get_air_quality_dataset
+from irvine.air_quality.aq_etl import aq_train_test_split, find_derivatives, get_air_quality_dataset
 from irvine.air_quality.lstm_model import (
     LSTM,
     ModelType,
@@ -83,29 +81,6 @@ def main() -> None:
         results[name] = train_and_evaluate_model(model, x_train, y_train, x_test, y_test)
 
     report(results)
-
-
-@beartype
-def aq_train_test_split(
-    x: pd.DataFrame,
-    y: NDArray[np.float64],
-    test_size: float = 0.2,
-    seed: int = 42,
-) -> tuple[NDArray[np.float64], NDArray[np.float64], NDArray[np.float64], NDArray[np.float64]]:
-    """This is simply a type safe wrapper of the familiar sklearn function."""
-    kwargs = {"test_size": test_size, "random_state": seed}
-
-    x_train, x_test, y_train, y_test = train_test_split(x, y, shuffle=False, **kwargs)
-
-    assert isinstance(x_train, pd.DataFrame)
-    assert isinstance(x_test, pd.DataFrame)
-
-    return (
-        x_train.to_numpy(),
-        x_test.to_numpy(),
-        pd.Series(y_train).to_numpy(),
-        pd.Series(y_test).to_numpy(),
-    )
 
 
 ModelTrainer = Callable[
